requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0

# Data handling and visualization
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.65.0

# Configuration and utilities
pyyaml>=6.0

# For training monitoring (optional)
tensorboard>=2.12.0

# For model evaluation
scikit-learn>=1.0.0

# For data augmentation (optional advanced features)
albumentations>=1.3.0

# Jupyter support (for notebooks)
jupyter>=1.0.0
ipykernel>=6.20.0

# Code quality (development)
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0

